Rename model folders and consolidate conversion scripts
Folder renames:
  models/default/          -> models/ACG/
  models/MHs/              -> models/MCG/
  models/radio_background/ -> models/radio/

Removed scripts/convert_tf_to_pytorch.py (duplicate of
models/ACG/convert_v1_to_pytorch.py); deleted scripts/ package.

Updated all references in:
  emulator.py, properties.py, models/ACG/v1_pytorch.py,
  models/MCG/__init__.py, tests/test_main.py, tests/test_v3.py,
  pyproject.toml (per-file-ignores), .pre-commit-config.yaml (exclude)
