+++ src/py21cmemu/models/ACG/v1_model.py
