+++ src/py21cmemu/models/MCG/lstm_emulator_constants.npz
