+++ src/py21cmemu/models/MCG/lstm_model.py
