+++ src/py21cmemu/models/MCG/score_model.py
