--- src/py21cmemu/scripts/__init__.py
@@ -1,1 +1,0 @@
-"""Utility scripts bundled with py21cmemu."""
